Add special detection of OTF locations
Adds a special check for bib location os and "none" (which appears
sometimes for OTF bibs). When filtering out non-research bibs, we should
consider bibs with 'os' or 'none' as location as Research so that they
are processed by discovery-api-indexer, which may choose to suppress
(i.e. delete) the record if it's actually an OTF record. This fixes an
issue where OTF bibs sneak into the index (because when they arrive they
have 0 items, qualifying them as Research) but are then not removed ...

https://jira.nypl.org/browse/SCC-2888