Fix bug identifying branch bibs by location
One of the checks employed to short circuit the indexing process
(and delete the bib) is a check on the bib's locations. If any of the
locations' collectionType properties contains 'Branch' and only 'Branch'
then that location is a Branch location and the bib is classified as
Branch. The check was wrong as it inadvertantly considered locations
with collectionTypes ['Branch', 'Research'] as being _only_ Branch.

This is an example bib affected by the bug: b21931073 . After
deployment, reindexing that bib should cause it to reappear.