Fix bug in statement grouping around deletions
Fix bug in latest refactor of code to group statements into
discovery-store-model classes wrapping statements representing bibs,
items, and holdings. The bug was that the new code overlooked that
deleted bibs will not be represented by an rdfs:type statement - instead
only including a nypl:suppressed statement to indicate they should be
deleted. This oversight caused the grouping code to fail to identify the
set of "bib statements", leading to a fatal runtime error. The fix
changes the way that bib statements are identified by identifying them
as the set of statments having `subject_id` equal to the bib's prefixed
id.